Change enforce_csrf from method to function
When it is necessary for a sublcass of TokenAuthentication to utilize
CSRF token checking in some cases, it was previously necessary to either
copy/paste the code from SessionAuthentication#enforce_csrf into that
subclass or create an instance of SessionAuthentication in order to
delegate to that method (which is ugly and feels fragile).

Making enforce_csrf a regular function makes it clear that it will never
depend on any instance state and can be invoked from any code that needs
it.